requirements.txt
fastapi
uvicorn
PyPDF2
langchain
google-generativeai
faiss-cpu
python-dotenv
fastapi[all]
